requirements.txt
Flask>=2.2
Pillow>=10.0
numpy>=1.25
scipy>=1.10 